Crate exhaust

This crate provides the Exhaust trait and derive macro, which allow iterating over all values of a given type.

Package features

All features are enabled by default. If you set default-features = false, exhaust becomes no_std compatible. The alloc and std features add Exhaust implementations for the corresponding standard library crates.

Modules

iteration

Assistance for implementing exhaustive iterators.

Structs

Iter

Iterator over all values of any type that implements Exhaust.

Traits

Exhaust

Types that can be exhaustively iterated. That is, an iterator is available which produces every possible value of this type.

Derive Macros

Exhaust

Derive macro generating an impl of the trait Exhaust.